requirements.txt
opencv-python
numpy
torch
torchvision
matplotlib
scipy
Pillow
tensorboard
ipykernel
tqdm
scikit-learn
imageio
pandas
accelerate
diffusers
Flask
albumentations
transformers